add classes for unit and category
