+++ unit_converter/unit_converter/lib/converter_route.dart
@@ -1,8 +1,12 @@
 import 'dart:core';
 import 'package:flutter/material.dart';
 
+import 'category.dart';
+import 'unit.dart';
+
 class ConverterPage extends StatefulWidget {
-  ConverterPage({Key key}) : super(key: key);
+  final Category category;
+  ConverterPage({Key key, this.category}) : super(key: key);
 
   @override
   _ConverterPageState createState() => new _ConverterPageState();
@@ -17,25 +21,34 @@ class _ConverterPageState extends State<ConverterPage> {
   @override
   Widget build(BuildContext context) {
     List<DropdownMenuItem> units = [];
-    units.add(new DropdownMenuItem(
-      value: 'value',
-      child: new Container(),
-    ));
+    for (Unit unit in widget.category.units) {
+      units.add(new DropdownMenuItem(
+        value: unit.name,
+        child: new Container(),
+      ));
+    }
+
+    // This is the widget that accepts text input. In this case, it accepts
+    // numbers. You can read more about it here: https://flutter.io/text-input
+    Widget input = new Container(
+        width: 300.0,
+        child: new TextField(
+          keyboardType: TextInputType.number,
+          onChanged: null,
+        ),
+    );
+
+    // This is the dropdown from where you can select unit types
     Widget dropdown = new DropdownButton(
       items: units,
       onChanged: updateConversion,
     );
+
     Widget convertFrom = new Container(
       color: Colors.pink,
       child: new Row(
         children: <Widget>[
-          new Container(
-              height: 50.0,
-              width: 300.0,
-              child: new TextField(
-                keyboardType: TextInputType.number,
-                onChanged: null,
-              )),
+          input,
           dropdown,
         ],
       ),
@@ -52,7 +65,6 @@ class _ConverterPageState extends State<ConverterPage> {
     );
 
     return new Container(
-      color: Colors.pink,
       child: new Column(
         crossAxisAlignment: CrossAxisAlignment.stretch,
         children: <Widget>[
