+++ unit_converter/unit_converter/lib/unit.dart
@@ -1,0 +1,7 @@
+class Unit {
+  String name;
+  double conversion;
+  String description;
+
+  Unit(this.name, this.conversion, this.description);
+}
